flixed_django/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.3.0
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
gunicorn==22.0.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==24.0
psycopg==3.2.4
psycopg-binary==3.2.4
PyJWT==1.7.1
python-dotenv==1.0.1
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tmdbsimple==2.9.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.7
